Rename synchronize_observer to synchronize_ranks and refactor moving average observers
- Rename Observer.synchronize_observer() -> synchronize_ranks() for clarity
- Remove get_current_min_max abstraction from MovingAverageObserverBase
- Each moving average observer now implements _update_statistics directly
- MovingAverageObserverBase is now a simpler mixin providing avg_constant and _lerp
- Update all test references to use synchronize_ranks()
- Remove unused imports (abstractmethod, MinMaxTuple) from moving_base.py

This makes the moving average observer pattern consistent with the base observer
pattern where each observer implements _update_statistics with its own logic.
